Implementação
Criação da classes de mapeamento do JSON de retorno da API

Criação da classe ApiServive que contém toda configuração do nosso serviço utilizando Retrofit e Moshi.

Na interface NYTServices foi ajustada a função que será utilizada pelo serviço para fazer a requisição na API

Foram adicionados artefatos no build.gradle das bibliotecas que serão utilizadas no projeto

Foram adicionadas algumas informações no arquivo de configuração gradle.properties, como url_base e api_key

Foi criada a configuração de buildTypes, para utilizar as informações que foram adicionadas no gradle.properties

Foi necessário adicionar o bloco compileOptions para compatibilidade com Java 8
