+++ app/src/main/java/com/example/newyorktimesbooks/data/response/BookBodyResponse.kt
@@ -1,0 +1,7 @@
+package com.example.newyorktimesbooks.data.response
+
+import com.squareup.moshi.Json
+
+data class BookBodyResponse(
+    @Json(name = "results") val bookResults: List<BookResultsResponse>
+)
