+++ app/src/main/java/com/example/newyorktimesbooks/data/response/BookResultsResponse.kt
@@ -1,0 +1,7 @@
+package com.example.newyorktimesbooks.data.response
+
+import com.squareup.moshi.Json
+
+data class BookResultsResponse(
+    @Json(name = "book_details") val details: List<BookDetailsResponse>
+)
