+++ app/src/main/java/com/example/newyorktimesbooks/data/response/BoookDetailsResponse.kt
@@ -1,0 +1,7 @@
+package com.example.newyorktimesbooks.data.response
+
+data class BookDetailsResponse(
+    val title: String,
+    val description: String,
+    val author: String
+)
