+++ app/src/main/java/com/example/newyorktimesbooks/data/service/ApiService.kt
@@ -1,0 +1,29 @@
+package com.example.newyorktimesbooks.data.service
+
+import com.example.newyorktimesbooks.BuildConfig
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+class ApiService {
+
+    private val retrofit: Retrofit
+
+    init {
+        val moshi = Moshi.Builder()
+            .add(KotlinJsonAdapterFactory())
+            .build()
+
+        retrofit = Retrofit.Builder()
+            .baseUrl(BuildConfig.URL_BASE)
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .addCallAdapterFactory(CoroutineCallAdapterFactory())
+            .build()
+    }
+
+    fun getService(): NYTServices {
+        return retrofit.create(NYTServices::class.java)
+    }
+}
