+++ app/src/main/java/com/example/newyorktimesbooks/data/service/NYTServices.kt
@@ -1,0 +1,15 @@
+package com.example.newyorktimesbooks.data.service
+
+import com.example.newyorktimesbooks.BuildConfig
+import com.example.newyorktimesbooks.data.response.BookBodyResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface NYTServices {
+
+    @GET("lists.json")
+    suspend fun getBooks(
+        @Query("api-key") apiKey: String = BuildConfig.API_KEY,
+        @Query("list") list: String = "hardcover-fiction"
+    ): List<BookBodyResponse>
+}
