+++ build.gradle
@@ -11,7 +11,10 @@ buildscript {
         ext_junit_version = '1.1.1'
         espresso_core_version = '3.2.0'
         lifecycle_version = "2.2.0"
-       retrofit_version = "2.9.0"
+        coroutine_version = "1.3.3"
+        retrofit_version = "2.9.0"
+        retrofit_coroutines_adapter_version = "0.9.2"
+        moshi_version = '1.9.2'
     }
     repositories {
         google()
