+++ gradle.properties
@@ -19,3 +19,5 @@ android.useAndroidX=true
 android.enableJetifier=true
 # Kotlin code style for this project: "official" or "obsolete":
 kotlin.code.style=official
+url_base="https://api.nytimes.com/svc/books/v3/"
+api_key="aGIJ2boHQbBUR2s8GSjtJ24839FmoIKV"
